+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - `Innmind\Html\Node\Document` no longer implement any interface
 - `Innmind\Html\Reader\Reader` now return an `Innmind\Immutable\Attempt`
 - `Innmind\Html\Visitor\Elements` now return a `Innmind\Immutable\Sequence`
+- `Innmind\Html\Reader\Reader` has been renamed `Innmind\Html\Reader`
 
 ### Removed